nsq_to_file: cleanup
